#61 moved some logic from core to trigger module
