+++ trigger-bot/src/main/java/ru/ustits/colleague/triggers/tools/LastTrigger.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.triggers;
+package ru.ustits.colleague.triggers.tools;
 
 import java.util.Collections;
 import java.util.List;
@@ -6,16 +6,15 @@ import java.util.List;
 /**
  * @author ustits
  */
-public final class FirstTrigger implements ProcessingStrategy {
+public final class LastTrigger implements ProcessingStrategy {
 
   @Override
   public List<String> process(final List<String> messages) {
     if (messages.isEmpty()) {
       return messages;
     } else {
-      final String first = messages.get(0);
+      final String first = messages.get(messages.size() - 1);
       return Collections.singletonList(first);
     }
   }
-
 }
