+++ trigger-bot/src/main/java/ru/ustits/colleague/triggers/tools/ProcessingStrategy.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.triggers;
+package ru.ustits.colleague.triggers.tools;
 
 import java.util.List;
