+++ trigger-bot/src/main/java/ru/ustits/colleague/triggers/tools/RandomTrigger.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.triggers;
+package ru.ustits.colleague.triggers.tools;
 
 import java.util.Collections;
 import java.util.List;
