+++ trigger-bot/src/main/java/ru/ustits/colleague/triggers/tools/TriggerProcessor.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.triggers;
+package ru.ustits.colleague.triggers.tools;
 
 import ru.ustits.colleague.repositories.records.TriggerRecord;
