+++ trigger-bot/src/test/java/ru/ustits/colleague/triggers/tools/AllTriggersTest.java
@@ -1,4 +1,4 @@
-package ru.ustits.colleague.tools.triggers;
+package ru.ustits.colleague.triggers.tools;
 
 import org.junit.Test;
